# doc/openapi.yaml
openapi: 3.0.3
info:
  title: cix-server API
  version: v1
  description: |
    HTTP API for the `cix-server` semantic code-index daemon.

    The wire format is byte-stable across server versions; the `api_version`
    field in `/api/v1/status` ticks independently from `server_version` when
    a backwards-incompatible change lands.

    ## Authentication

    Two parallel auth paths back the same identity model:

    - **Browsers / dashboard**: `POST /api/v1/auth/login` with email +
      password issues an HttpOnly cookie `cix_session`. The cookie is sent
      automatically on subsequent same-origin requests. Sessions roll
      forward 14 days from the last request.
    - **CLI / SDK**: `Authorization: Bearer <api_key>` where the key was
      issued by `POST /api/v1/api-keys`. Keys are owner-scoped and can be
      revoked individually without affecting other clients.

    Public endpoints (no auth required): `GET /health`, `GET /docs`,
    `GET /openapi.yaml`, `GET /api/v1/auth/bootstrap-status`,
    `POST /api/v1/auth/login`. Setting `CIX_AUTH_DISABLED=true` skips the
    check on every endpoint (development only — a warning is logged).

    When the server starts with an empty users table it requires
    `CIX_BOOTSTRAP_ADMIN_EMAIL` + `CIX_BOOTSTRAP_ADMIN_PASSWORD` to seed
    the first admin (forced to change password on first login). A legacy
    `CIX_API_KEY` set on a fresh database is imported as a single
    `env-bootstrap` API key owned by that admin.

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/runtime-config:
    get:
      operationId: getRuntimeConfig
      tags: [admin]
      summary: Read effective runtime config (admin only)
      description: |
        Returns the resolved runtime config (DB row → env → recommended) the
        sidecar is currently configured against, plus a `source` map labelling
        each field's origin so the dashboard can render the "DB" / "Env" /
        "Recommended" pill next to every value.
      responses:
        "200":
          description: Effective runtime config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      operationId: putRuntimeConfig
      tags: [admin]
      summary: Save runtime config overrides (admin only)
      description: |
        Replaces the runtime_settings row. Fields omitted from the request
        clear their override (= fall back to env / recommended on next Get).
        Does NOT restart the sidecar — the dashboard issues a separate
        POST /admin/sidecar/restart after a successful PUT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeConfigUpdate"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /api/v1/admin/sidecar/restart:
    post:
      operationId: restartSidecar
      tags: [admin]
      summary: Restart the llama-server sidecar (admin only)
      description: |
        Drains the embedding queue (30s timeout), terminates the current
        child process, and respawns with the latest runtime config.
        Returns 202 immediately; poll GET /admin/sidecar/status to observe
        the running → restarting → running transition. In-flight indexing
        batches at the moment of restart will fail with ErrSupervisor and
        must be re-driven by the operator (`cix reindex <project>`).
      responses:
        "202":
          description: Restart accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestartAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          description: Embeddings disabled at boot — restart is a no-op
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/admin/sidecar/status:
    get:
      operationId: getSidecarStatus
      tags: [admin]
      summary: Sidecar process status (admin only)
      responses:
        "200":
          description: Status snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SidecarStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/models:
    get:
      operationId: listModels
      tags: [admin]
      summary: List GGUF model files cached on disk (admin only)
      description: |
        Walks `CIX_GGUF_CACHE_DIR` and returns one entry per .gguf file. Used
        by the dashboard's embedding-model picker so admins don't have to
        type HF repo IDs by hand. Empty list when the cache is empty —
        dashboard falls back to a free-text path input in that case.
      responses:
        "200":
          description: Cached model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key passed as `Authorization: Bearer <CIX_API_KEY>`"
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unprocessable:
      description: Malformed request body or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Authenticated, but lacks the required role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
...
    RuntimeConfig:
      type: object
      required:
        - embedding_model
        - llama_ctx_size
        - llama_n_gpu_layers
        - llama_n_threads
        - max_embedding_concurrency
        - llama_batch_size
        - source
      properties:
        embedding_model:
          type: string
          description: HF repo ID or absolute filesystem path to a .gguf file.
        llama_ctx_size:
          type: integer
          minimum: 1
        llama_n_gpu_layers:
          type: integer
          description: -1 = all layers (Metal/CUDA), 0 = CPU only.
        llama_n_threads:
          type: integer
          minimum: 0
          description: 0 = let llama-server auto-detect.
        max_embedding_concurrency:
          type: integer
          minimum: 1
        llama_batch_size:
          type: integer
          minimum: 1
        source:
          type: object
          additionalProperties:
            type: string
            enum: [db, env, recommended]
          description: |
            Per-field origin label so the dashboard can render a "DB" /
            "Env" / "Recommended" pill next to each value. Keys match the
            other field names: `embedding_model`, `llama_ctx_size`, ...
        recommended:
          $ref: "#/components/schemas/RuntimeConfigRecommended"
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: When the runtime_settings row was last written, or null when only env/recommended are in effect.
        updated_by:
          type: string
          nullable: true
          description: Who issued the last PUT, captured from the active session.
    RuntimeConfigRecommended:
      type: object
      required:
        - embedding_model
        - llama_ctx_size
        - llama_n_gpu_layers
        - llama_n_threads
        - max_embedding_concurrency
        - llama_batch_size
      properties:
        embedding_model: { type: string }
        llama_ctx_size: { type: integer }
        llama_n_gpu_layers: { type: integer }
        llama_n_threads: { type: integer }
        max_embedding_concurrency: { type: integer }
        llama_batch_size: { type: integer }
    RuntimeConfigUpdate:
      type: object
      description: |
        All fields optional. Send a value to set/replace the override for
        that field, send `""` (string fields) or `0` (numeric fields) to
        CLEAR the override (next read falls back to env / recommended).
        Omitted fields keep their current value.
      properties:
        embedding_model:
          type: string
          nullable: true
        llama_ctx_size:
          type: integer
          nullable: true
        llama_n_gpu_layers:
          type: integer
          nullable: true
        llama_n_threads:
          type: integer
          nullable: true
        max_embedding_concurrency:
          type: integer
          nullable: true
        llama_batch_size:
          type: integer
          nullable: true
    SidecarStatus:
      type: object
      required: [state, ready, in_flight]
      properties:
        state:
          type: string
          enum: [running, starting, restarting, failed, disabled]
        pid:
          type: integer
          minimum: 0
          description: 0 when no child process is alive (failed / disabled).
        uptime_seconds:
          type: integer
          minimum: 0
        model:
          type: string
        ready:
          type: boolean
        last_error:
          type: string
        in_flight:
          type: integer
          minimum: 0
          description: Embedding queue depth at the moment of sampling.
        restart_in_flight:
          type: boolean
          description: True between accept of POST /sidecar/restart and respawn completion.
    RestartAccepted:
      type: object
      required: [restart_id]
      properties:
        restart_id:
          type: string
          description: Opaque ID; future versions may expose per-restart progress under this id.
    ModelEntry:
      type: object
      required: [id, path, size_bytes]
      properties:
        id:
          type: string
          description: HF repo ID derived from the cache directory name (e.g. owner/model).
        path:
          type: string
          description: Absolute path to the .gguf file on disk.
        size_bytes:
          type: integer
          format: int64
          minimum: 0
    ModelList:
      type: object
      required: [models, cache_dir]
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelEntry"
        cache_dir:
          type: string
          description: The CIX_GGUF_CACHE_DIR that was scanned. Empty list with non-empty cache_dir = no .gguf files found.
...